// app/community/types.ts
export interface LocationFilterType {
  lat: number;
  lng: number;
  radius: number;
}